refactor : 텍스트의 이미지 s3에 저장하는 함수 분리
