+++ libs/components-seller/src/lib/goods-regist/GoodsRegistForm.tsx
@@ -20,7 +20,7 @@ import {
   RegistGoodsDto,
 } from '@project-lc/shared-types';
 import { goodsRegistStore } from '@project-lc/stores';
-import { s3 } from '@project-lc/utils-s3';
+import { saveContentsImageToS3 } from '@project-lc/utils-frontend';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
@@ -37,22 +37,6 @@ import GoodsRegistMemo from './GoodsRegistMemo';
 import GoodsRegistPictures from './GoodsRegistPictures';
 import GoodsRegistShippingPolicy from './GoodsRegistShippingPolicy';
 
-// 상품 사진, 상세설명 이미지를 s3에 업로드 -> url 리턴
-export async function uploadGoodsImageToS3(
-  imageFile: { file: File | Buffer; filename: string; id: number; contentType: string },
-  userMail: string,
-): Promise<string> {
-  const { file, filename, contentType } = imageFile;
-  return s3.s3UploadImage({
-    isPublic: true,
-    file,
-    filename,
-    ContentType: contentType,
-    userMail,
-    type: 'goods',
-  });
-}
-
 // options 에 default_option설정
 export function addGoodsOptionInfo(
   options: Omit<GoodsOptionDto, 'default_option'>[],
@@ -61,40 +45,6 @@ export function addGoodsOptionInfo(
     ...opt,
     default_option: index === 0 ? ('y' as const) : ('n' as const),
   }));
-}
-
-// 상품 '상세설명' contents에서 이미지 s3에 업로드 후 src url 변경
-// 상세설명 이미지는data:image 형태로 들어있음
-export async function saveContentsImageToS3(
-  contents: string,
-  userMail: string,
-): Promise<string> {
-  const parser = new DOMParser();
-  const dom = parser.parseFromString(contents, 'text/html'); // textWithImages -> getValues('contents')
-  const imageTags = Array.from(dom.querySelectorAll('img'));
-
-  await Promise.all(
-    imageTags.map(async (tag, index) => {
-      const src = tag.getAttribute('src');
-      const name = tag.dataset.fileName || '';
-      if (src && src.slice(0, 4) !== 'http') {
-        const imageBuffer = Buffer.from(
-          src.replace(/^data:image\/\w+;base64,/, ''),
-          'base64',
-        );
-        const fileType = src.substring('data:'.length, src.indexOf(';base64'));
-        const url = await uploadGoodsImageToS3(
-          { file: imageBuffer, filename: name, id: index, contentType: fileType },
-          userMail,
-        );
-        // img src 바꾸기
-        tag.setAttribute('src', url);
-      }
-    }),
-  );
-
-  const contentsBody = dom.body.innerHTML;
-  return contentsBody;
 }
 
 /** 상품 등록 폼 컴포넌트 */
